pyproject.toml
[project]
name = "investtipsdaily"
version = "0.1.0"
description = "TipsAI — Bot inteligente do grupo Invest Tips Daily"
requires-python = ">=3.10,<3.13"
dependencies = [
    "python-telegram-bot>=21.0",
    "anthropic>=0.40.0",
    "beautifulsoup4>=4.12.0",
    "sentence-transformers>=3.0.0",
    "chromadb>=1.0.0",
    "python-dotenv>=1.0.0",
    "lxml>=5.0.0",
    "ddgs>=7.0.0",
    "openai-whisper>=20231117",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
